requirements.txt
numpy
scipy
pillow
matplotlib
opencv-python